🚚 move allocated time to dto
