+++ backend/api/src/presentation/graphql/dto/mod.rs
@@ -6,3 +6,6 @@ pub use payment::Payment;
 
 mod language;
 pub use language::Language;
+
+mod user_profile;
+pub use user_profile::AllocatedTime;
