--- backend/common/domain/src/user_profile/allocated_time.rs
@@ -1,10 +1,0 @@
-use juniper::GraphQLEnum;
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, GraphQLEnum)]
-pub enum AllocatedTime {
-	None,
-	LessThanOneDay,
-	OneToThreeDays,
-	MoreThanThreeDays,
-}
