--- backend/common/domain/src/user_profile/mod.rs
@@ -1,2 +1,0 @@
-mod allocated_time;
-pub use allocated_time::AllocatedTime;
